+++ contracts/interfaces/IEtherbase.sol
@@ -25,5 +25,5 @@ pragma solidity ^0.8.0;
 interface IEtherbase {
     receive() external payable;
     function retrieve(address payable receiver) external;
-    function partiallyRetrieve(address payable receiver, uint amount) external;
+    function partiallyRetrieve(address payable receiver, uint256 amount) external;
 }
